Rename types file
